pyproject.toml
[tool.poetry]
name = "Deck"
version = "0.1.0"
description = "Deck, a CLI that creates reproducible Kubernetes environments for development and testing"
authors = ["Michael Schilonka <michael@unikube.io>"]
readme = "README.md"
homepage = "https://getdeck.dev"
repository = "https://github.com/Schille/deck"
documentation = "https://getdeck.dev"
keywords = [
    "Kubernetes",
    "Development",
    "Cloud-native"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux"
]

[tool.poetry.dependencies]
python = "^3.10.0"
kubernetes = "^23.3.0"
docker = "^5.0.3"
pydantic = "^1.9.0"
semantic-version = "^2.9.0"
GitPython = "^3.1.27"


[tool.poetry.dev-dependencies]
flake8-bugbear = "^22.1.11"
black = "^22.1.0"
flake8 = "^4.0.1"
flake8-black = "^0.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
deck = "deck.__main__:main"

[tool.poetry.urls]
issues = "https://github.com/Schille/deck/issues"
